fix for merger script

+++ scripts/merger.py
@@ -2,8 +2,6 @@ import os
 import re
 import sys
 
-output_dir = "D:/git-workspace/WonderDev/toto/generated_src"
-output_file = "D:/git-workspace/WonderDev/toto/generated_src/codingame.cpp"
 
 excluded_files = ["inGameMain.txt", "Algo.cpp", "Algo.h", "Evaluation.cpp", "Evaluation.h"]
 
@@ -13,27 +11,39 @@ processed_files = []
 
 
 def run():
+
     working_dir = sys.argv[1]
 
-    if not os.path.exists(output_dir):
-        os.makedirs(output_dir)
+    if not os.path.exists(working_dir):
+        print "Path does not exist : " + working_dir
+        return
+    else:
+        src_dir = os.path.join(working_dir, "src")
+        if not os.path.exists(src_dir):
+            print "Path does not exist : " + src_dir
+            return
+        output_dir = os.path.join(working_dir, "generated_src")
+        output_file = os.path.join(output_dir, "codingame.cpp")
 
-    # clean
-    with open(output_file, "w") as output:
-        output.write("")
-    process_file(os.path.join(working_dir, "inGameMain.txt"), False)
-    process_dir(working_dir)
+        if not os.path.exists(output_dir):
+            os.makedirs(output_dir)
+
+        # clean
+        with open(output_file, "w") as output:
+            output.write("")
+        process_file(os.path.join(src_dir, "inGameMain.txt"), False, output_file)
+        process_dir(src_dir, output_file)
 
 
-def process_dir(working_dir):
-    files = [f for f in os.listdir(working_dir) if os.path.isfile(os.path.join(working_dir, f))]
+def process_dir(src_dir, output_file):
+    files = [f for f in os.listdir(src_dir) if os.path.isfile(os.path.join(src_dir, f))]
 
     for file in files:
         if file not in excluded_files:
-            process_file(os.path.join(working_dir, file), False)
+            process_file(os.path.join(src_dir, file), False, output_file)
 
 
-def process_file(file, is_included):
+def process_file(file, is_included, output_file):
     if file.endswith(".h") and not is_included:
         return
     if file.endswith("main.cpp"):
@@ -59,7 +69,7 @@ def process_file(file, is_included):
                     if m:
                         included_file = os.path.join(os.path.dirname(file), m.group(1))
                         print "Found a file to include : " + included_file
-                        process_file(included_file, True)
+                        process_file(included_file, True, output_file)
                     else:
                         print "warning : " + line
                 elif line.startswith("#pragma once"):
@@ -79,7 +89,7 @@ def process_file(file, is_included):
 
 if __name__ == '__main__':
     if len(sys.argv) < 2:
-        print "Usage : merger.py path_to_sources_files"
-        print "Example : merger.py D:/git-workspace/WonderDev/toto/src"
+        print "Usage : merger.py path_to_project"
+        print "Example : merger.py D:/git-workspace/toto"
     else:
         run()
